Move mocking of exchange response for history creation test to utils
